modify(#14):패치 요청 보낼 때 alert 창 띄우기, hook 이름 변경,아이템 key값 추가
